requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
langchain-openai
chromadb
langchain_community
sentence_transformers
langchain_huggingface
faiss-cpu
beautifulsoup4
langchain-google-genai
streamlit